Switch to MediaElement.js!
While the native HTML5 audio player is a good target, we'd end up
writing our own JS/CSS to have a consistent cross-browser look and feel,
which is important to have in a platform used by so many.  There's also
accessibility concerns around the implementation of native players.

MediaElement.js is a cross-browser solution to providing a consistent UI
around the HTML5 audio/video spec, and it also has support for shimming
in Flash/Siverlight fallbacks for browsers that don't natively support
HTML5 audio/video.  As mentioned above, it also comes out of the box
with accessibility in mind, which is super important to have.