fix: Fix controls when end time is unset
Previously, without an end time set,
there were issues with the mute button and the progress slider.

Fix this by improving how the endTime is calculated in the frontend,
and checking if endTime is set before running any of the custom logic
for clipping the play based on start and end times.

Private-ref: https://tasks.opencraft.com/browse/BB-9935