fix: improve delete transcript UI
Change the process to delete a transcript file.
Previously it was a checkbox to delete on save.
Now it's a button that immediately deletes the file.

This was determined to be better
because it makes it clear that it's a deletion action,
and it avoids the deletion happening as part of saving.
See https://github.com/open-craft/xblock-audio/pull/6#discussion_r2268785655
for background discussion.

Private-ref: https://tasks.opencraft.com/browse/BB-9935