feat!: remove transcript file field
The transcript file field is problematic,
as it stores the file in a location
that is not included in course exports.
This means if a transcript file is uploaded,
then the course exported and imported into another Open edX instance,
the file is lost.

The field is also not currently in use as far as we're aware,
so we can simply remove the field.

This still allows timed transcripts, via the transcript url field.

In future if we need support, we will need to consider a better way to handle file uploads.
For example, the core Video XBlock handles handout file uploads
by directly uploading them to the course Files assets.
This isn't ideal though.
There are also other methods being considered for future development,
so we may have a better supported method in the future.

Private-ref: https://tasks.opencraft.com/browse/BB-10249